veteran-ai-spark/requirements.txt
flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0
openai==1.52.0
pinecone-client==4.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2
numpy==1.24.3
scikit-learn==1.3.0
tiktoken==0.5.1
pydantic==2.5.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pytest==7.4.3
pytest-cov==4.1.0
markdown==3.5.1